Update length-related operations
* remove each unnecessary length method
* getArrayLength and lengthES6 have been renamed
* return type of toLength is changed to uint64_t
* use createListFromArrayLike for more cases

+++ src/runtime/Value.h
@@ -202,7 +202,7 @@ public:
     inline bool toBoolean(ExecutionState& ec) const; // $7.1.2 ToBoolean
     double toNumber(ExecutionState& ec) const; // $7.1.3 ToNumber
     double toInteger(ExecutionState& ec) const; // $7.1.4 ToInteger
-    double toLength(ExecutionState& ec) const;
+    uint64_t toLength(ExecutionState& ec) const;
     int32_t toInt32(ExecutionState& ec) const; // $7.1.5 ToInt32
     uint32_t toUint32(ExecutionState& ec) const; // http://www.ecma-international.org/ecma-262/5.1/#sec-9.6
     String* toString(ExecutionState& ec) const // $7.1.12 ToString
